shifts from HFTask to Task removing common file
